requirements.txt
discord.py==2.3.2
python-dotenv==1.0.0
pytz==2024.1
colorlog==6.8.2  # For colored logging 